# swagger.json
openapi: 3.0.1
info:
  title: Api favorita Aiqfome
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/favorites:
    post:
      tags:
        - Produtos Favoritos
      summary: Adiciona um produto à lista de favoritos de um cliente
      description: Permite adicionar um produto à lista de favoritos de um cliente específico. Verifica se o produto existe na Fake Store API e se já não está favoritado pelo cliente.
      operationId: addFavorite
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteProductRequestDTO'
        required: true
      responses:
        '201':
          description: Produto adicionado aos favoritos com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteProductResponseDTO'
        '400':
          description: 'Requisição inválida: ID de produto inválido ou produto não encontrado na Fake Store API'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: 'Conflito: Produto já favoritado pelo cliente'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/products:
    get:
      tags:
        - Produtos
      summary: Lista todos os produtos
      description: Retorna uma lista de todos os produtos disponíveis, obtidos de uma API externa (Fake Store API).
      operationId: getAllProducts
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Erro interno ao buscar produtos na API externa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /api/v1/products/{id}:
    get:
      tags:
        - Produtos
      summary: Busca um produto por ID
      description: Retorna os detalhes de um produto específico, obtido de uma API externa (Fake Store API), usando seu ID.
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Produto encontrado e retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: 'Requisição inválida: ID de produto inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Produto não encontrado na API externa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Erro interno ao buscar o produto na API externa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /api/v1/favorites/client/{clientId}:
    get:
      tags:
        - Produtos Favoritos
      summary: Lista produtos favoritos de um cliente
      description: Retorna todos os produtos que um cliente específico favoritou.
      operationId: getFavoritesByClient
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de produtos favoritos retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteList'
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '409':
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /api/v1/favorites/{favoriteId}:
    delete:
      tags:
        - Produtos Favoritos
      summary: Remove um produto da lista de favoritos
      description: Remove um produto favorito de um cliente usando o ID do registro de favorito.
      operationId: removeFavorite
      parameters:
        - name: favoriteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Produto removido dos favoritos com sucesso (sem conteúdo na resposta)
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Falha de autenticacao
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Registro de favorito não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
components:
  schemas:
...
    ProductList:
      type: array
      items:
          $ref: '#/components/schemas/ProductDTO' 
    FavoriteList:
      type: array
      items:
          $ref: '#/components/schemas/FavoriteProductResponseDTO' 
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
        details:
          type: string
...
    FavoriteProductRequestDTO:
      required:
        - clientId
        - productId
      type: object
      properties:
        productId:
          type: integer
          format: int64
        clientId:
          type: string
          format: uuid
    FavoriteProductResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        titulo:
          type: string
        imagem:
          type: string
        preco:
          type: number
        review:
          type: number
        clientId:
          type: string
          format: uuid
        productId:
          type: integer
          format: int64
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        category:
          type: string
        image:
          type: string
        rating:
          $ref: '#/components/schemas/RatingDTO'
    RatingDTO:
      type: object
      properties:
        rate:
          type: number
          format: double
        count:
          type: integer
          format: int64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT